+++ lib/rules/sotd/pp.js
@@ -70,22 +70,10 @@ function buildWanted(groups, sr, ppLink) {
  */
 function findPP(candidates, sr, isIGDeliverable) {
     let pp = null;
-    const groups = [];
-    const jointRegex = new RegExp(
-        'This document was (?:produced|published) by the (.+? Working Group|Technical Architecture Group|Advisory Board)' +
-            ' and the (.+? Working Group|Technical Architecture Group|Advisory Board)'
-    );
-    Array.prototype.some.call(candidates, p => {
-        const text = sr.norm(p.textContent);
-        if (jointRegex.test(text)) {
-            const matches = text.match(jointRegex);
-            groups.push(matches[1]);
-            groups.push(matches[2]);
-            sr.warning(self, 'joint-publication');
-            return true;
-        }
-    });
-    const wanted = buildWanted(groups, sr, isIGDeliverable);
+    const delivererGroups = sr.getDelivererGroups();
+    if (delivererGroups.length > 1) sr.warning(self, 'joint-publication');
+
+    const wanted = buildWanted(delivererGroups, sr, isIGDeliverable);
     const expected = wanted.text;
     Array.prototype.some.call(candidates, p => {
         const text = sr.norm(p.textContent);
