+++ lib/rules/sotd/stability.js
@@ -1,7 +1,7 @@
 // SotD
 //  stability warning
 // <p>Publication as a Working Draft does not imply endorsement by W3C and its Members.</p>
-import { filter, getDelivererNames } from '../../util.js';
+import { filter } from '../../util.js';
 
 /**
  * @param candidates
@@ -15,9 +15,8 @@ function findSW(candidates, sr) {
         sr.config.longStatus === 'Group Draft Note'
     ) {
         // Find the sentence of 'Group Notes are not endorsed by W3C nor its Members.' or 'This Group Note is endorsed by the @@ Group, but is not endorsed by W3C itself nor its Members.'
-        wanted = `(${sr.config.longStatus}s are not endorsed by W3C nor its Members|This ${sr.config.longStatus} is endorsed by the \\w+ [Working|Interest] Group, but is not endorsed by W3C itself nor its Members).`;
-        const groups = getDelivererNames(sr);
-        console.log('\ngroups: ', groups);
+        const groups = sr.getDelivererNames().join(' and the ');
+        wanted = `(${sr.config.longStatus}s are not endorsed by W3C nor its Members|This ${sr.config.longStatus} is endorsed by the ${groups}, but is not endorsed by W3C itself nor its Members).`;
     } else if (sr.config.longStatus === 'Statement') {
         wanted =
             'A W3C Statement is a specification that, after extensive consensus-building, is endorsed by W3C and its Members.';
