+++ lib/util.js
@@ -237,18 +237,3 @@ export const REC_TEXT = {
     SOTD_C_COR_ADD:
         'It includes candidate amendment(s)?, introducing substantive change(s)? and new feature(s)? since the previous Recommendation.',
 };
-
-export const getDelivererNames = function (sr) {
-    const sotd = sr.getSotDSection();
-    const delivererNamesRegex =
-        /This document was (?:produced|published) by the ([^Group]+ Working Group|[^Group]+ Interest Group|Technical Architecture Group|Advisory Board)( and the (.+? Working Group|.+? Interest Group|Technical Architecture Group|Advisory Board))?/;
-
-    const text = sr.norm(sotd.textContent);
-    const matches = text.match(delivererNamesRegex);
-    const groups = [];
-    if (matches) {
-        groups.push(matches[1]);
-        if (matches[2]) groups.push(matches[2]);
-    }
-    return groups;
-};
