+++ lib/validator.js
@@ -509,6 +509,22 @@ Specberus.prototype.getFeedbackDueDate = function () {
     return dates;
 };
 
+// Return array of group names, e.g. ['Internationalization Working Group', 'Technical Architecture Group']
+export const getDelivererNames = function () {
+    const sotd = this.getSotDSection();
+    const delivererNamesRegex =
+        /This document was (?:produced|published) by the (\w+ Working Group|\w+ Interest Group|Technical Architecture Group|Advisory Board)( and the (\w+ Working Group|\w+ Interest Group|Technical Architecture Group|Advisory Board))?/;
+
+    const text = this.norm(sotd.textContent);
+    const matches = text.match(delivererNamesRegex);
+    const groups = [];
+    if (matches) {
+        groups.push(matches[1]);
+        if (matches[3]) groups.push(matches[3]);
+    }
+    return groups;
+};
+
 /**
  * getDelivererGroups get deliverers groupNames and types
  *
